+++ src/test/java/ch/aaap/harvestclient/impl/estimate/EstimatesApiCreateTest.java
@@ -75,13 +75,13 @@ class EstimatesApiCreateTest {
     @Test
     void createWithLineItems() {
 
-        long quantity = 20;
+        double quantity = 20;
         double unitPrice = 30;
         String firstName = "test First";
         Estimate creationInfo = ImmutableEstimate.builder()
                 .client(clientReference)
                 .addEstimateItem(ImmutableEstimateItem.builder()
-                        .quantity(10L)
+                        .quantity(10.)
                         .unitPrice(1.)
                         .kind(kind)
                         .build())
